py/requirements.txt
grpcio-tools==1.56.0
numpy==1.25.1